Use ActiveSupport::Concern to add the hook to asset, work and collection models
