+++ app/models/work.rb
@@ -1,6 +1,8 @@
 class Work < Kithe::Work
   # will trigger automatic solr indexing in callbacks
 
+  include RecordPublishedAt
+
   if ScihistDigicoll::Env.lookup(:solr_indexing) == 'true'
     self.kithe_indexable_mapper = WorkIndexer.new
   end
@@ -39,8 +41,6 @@ class Work < Kithe::Work
     work.validates_presence_of :department
     work.validates_presence_of :rights
   end
-
-  before_validation :set_published_at
 
   attr_json :additional_title, :string, array: true, default: -> { [] }
   attr_json :external_id, Work::ExternalId.to_type, array: true, default: -> { [] }
@@ -214,11 +214,6 @@ class Work < Kithe::Work
   # sense to make a method to encapsulate it.
   def is_oral_history?
     genre && genre.include?("Oral histories")
-  end
-
-  private
-  def set_published_at
-    self.published_at = DateTime.now if published? && published_changed?
   end
 
 end
